Creating DugFactory
This change creates a DugFactory class which is responsible for
constructing all the necessary parts of the Dug system. It also does
some cleanup and organization.
