+++ src/dug/core/loaders/network_loader.py
