+++ src/dug/core/search.py
@@ -1,50 +1,12 @@
 import json
 import logging
-import os
-import sys
-import traceback
-from functools import partial
-from pathlib import Path
-from typing import Iterable
 
-import pluggy
-import redis
 import requests
 from elasticsearch import Elasticsearch
-from requests_cache import CachedSession
 
-from dug import annotate as anno
 from dug.config import Config
-from dug import hookspecs
-from dug import parsers
-from dug.loaders.filesystem_loader import load_from_filesystem
-from dug.loaders.network_loader import load_from_network
-from dug.parsers import DugElement, DugConcept, Parser
 
 logger = logging.getLogger('dug')
-stdout_log_handler = logging.StreamHandler(sys.stdout)
-formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-stdout_log_handler.setFormatter(formatter)
-logger.addHandler(stdout_log_handler)
-
-logging.getLogger("elasticsearch").setLevel(logging.WARNING)
-
-def get_plugin_manager() -> pluggy.PluginManager:
-    pm = pluggy.PluginManager("dug")
-    pm.add_hookspecs(hookspecs)
-    pm.load_setuptools_entrypoints("dug")
-    pm.register(parsers)
-    return pm
-
-
-class SearchException(Exception):
-    def __init__(self, message, details):
-        self.message = message
-        self.details = details
-
-
-class ParserNotFoundException(Exception):
-    ...
 
 
 class Search:
@@ -57,7 +19,11 @@ class Search:
          * disease->study
          * disease->phenotype->study
     """
-    def __init__(self, cfg: Config, indices=['concepts_index', 'variables_index', 'kg_index']):
+
+    def __init__(self, cfg: Config, indices=None):
+
+        if indices is None:
+            indices = ['concepts_index', 'variables_index', 'kg_index']
 
         self._cfg = cfg
         logger.debug(f"Connecting to elasticsearch host: {self._cfg.elastic_host} at port: {self._cfg.elastic_port}")
@@ -81,10 +47,7 @@ class Search:
                 details=f"connecting to host {self._cfg.elastic_host} and port {self._cfg.elastic_port}")
 
     def init_indices(self):
-        settings = {}
-
-        # kg_index
-        settings['kg_index'] = {
+        kg_index = {
             "settings": {
                 "number_of_shards": 1,
                 "number_of_replicas": 0
@@ -100,9 +63,7 @@ class Search:
                 }
             }
         }
-
-        # concepts_index
-        settings['concepts_index'] = {
+        concepts_index = {
             "settings": {
                 "index.mapping.coerce": "false",
                 "number_of_shards": 1,
@@ -130,9 +91,7 @@ class Search:
                 }
             }
         }
-
-        # variables_index
-        settings['variables_index'] = {
+        variables_index = {
             "settings": {
                 "index.mapping.coerce": "false",
                 "number_of_shards": 1,
@@ -157,7 +116,14 @@ class Search:
             }
         }
 
-        logger.info(f"creating indices: {self.indices}")
+        settings = {
+            'kg_index': kg_index,
+            'concepts_index': concepts_index,
+            'variables_index': variables_index,
+        }
+
+        logger.info(f"creating indices")
+        logger.debug(self.indices)
         for index in self.indices:
             try:
                 if self.es.indices.exists(index=index):
@@ -434,280 +400,7 @@ class Search:
             doc_id=unique_doc_id)
 
 
-class Crawler:
-    def __init__(self, crawl_file: str, parser: Parser, annotator,
-                 tranqlizer, tranql_queries,
-                 http_session, exclude_identifiers=[], element_type=None):
-
-        self.crawl_file = crawl_file
-        self.parser: Parser = parser
-        self.element_type = element_type
-        self.annotator = annotator
-        self.tranqlizer = tranqlizer
-        self.tranql_queries = tranql_queries
-        self.http_session = http_session
-        self.exclude_identifiers = exclude_identifiers
-        self.elements = []
-        self.concepts = {}
-        self.crawlspace = "crawl"
-
-    def make_crawlspace(self):
-        if not os.path.exists(self.crawlspace):
-            try:
-                os.makedirs(self.crawlspace)
-            except Exception as e:
-                print(f"-----------> {e}")
-                traceback.print_exc()
-
-    def crawl(self):
-
-        # Create directory for storing temporary results
-        self.make_crawlspace()
-
-        # Read in elements from parser
-        self.elements = self.parser(self.crawl_file)
-
-        # Optionally coerce all elements to be a specific type
-        for element in self.elements:
-            if isinstance(element, DugElement) and self.element_type is not None:
-                element.type = self.element_type
-
-        # Annotate elements
-        self.annotate_elements()
-
-        # Expand concepts
-        concept_file = open(f"{self.crawlspace}/concept_file.json", "w")
-        for concept_id, concept in self.concepts.items():
-            # Use TranQL queries to fetch knowledge graphs containing related but not synonymous biological terms
-            self.expand_concept(concept)
-
-            # Traverse identifiers to create single list of of search targets/synonyms for concept
-            concept.set_search_terms()
-
-            # Traverse kg answers to create list of optional search targets containing related concepts
-            concept.set_optional_terms()
-
-            # Remove duplicate search terms and optional search terms
-            concept.clean()
-
-            # Write concept out to a file
-            concept_file.write(f"{json.dumps(concept.get_searchable_dict(), indent=2)}")
-
-        # Close concept file
-        concept_file.close()
-
-    def annotate_elements(self):
-
-        # Open variable file for writing
-        variable_file = open(f"{self.crawlspace}/element_file.json", "w")
-
-        # Annotate elements/concepts and create new concepts based on the ontology identifiers returned
-        for element in self.elements:
-            # If element is actually a pre-loaded concept (e.g. TOPMed Tag), add that to list of concepts
-            if isinstance(element, DugConcept):
-                self.concepts[element.id] = element
-
-            # Annotate element with normalized ontology identifiers
-            self.annotate_element(element)
-            if isinstance(element, DugElement):
-                variable_file.write(f"{element}\n")
-
-        # Now that we have our concepts and elements fully annotated, we need to
-        # Make sure elements inherit the identifiers from their user-defined parent concepts
-        # E.g. TOPMedTag1 was annotated with HP:123 and MONDO:12.
-        # Each element assigned to TOPMedTag1 needs to be associated with those concepts as well
-        for element in self.elements:
-            # Skip user-defined concepts
-            if isinstance(element, DugConcept):
-                continue
-
-            # Associate identifiers from user-defined concepts (see example above)
-            # with child elements of those concepts
-            concepts_to_add = []
-            for concept_id, concept in element.concepts.items():
-                for ident_id, identifier in concept.identifiers.items():
-                    if ident_id not in element.concepts and ident_id in self.concepts:
-                        concepts_to_add.append(self.concepts[ident_id])
-
-            for concept_to_add in concepts_to_add:
-                element.add_concept(concept_to_add)
-
-        # Write elements out to file
-        variable_file.close()
-
-    def annotate_element(self, element):
-        # Annotate with a set of normalized ontology identifiers
-        identifiers = self.annotator.annotate(text=element.ml_ready_desc,
-                                              http_session=self.http_session)
-
-        # Each identifier then becomes a concept that links elements together
-        for identifier in identifiers:
-            if identifier.id not in self.concepts:
-                # Create concept for newly seen identifier
-                concept = DugConcept(concept_id=identifier.id,
-                                                       name=identifier.label,
-                                                       desc=identifier.description,
-                                                       concept_type=identifier.type)
-                # Add to list of concepts
-                self.concepts[identifier.id] = concept
-
-            # Add identifier to list of identifiers associated with concept
-            self.concepts[identifier.id].add_identifier(identifier)
-
-            # Create association between newly created concept and element
-            # (unless element is actually a user-defined concept)
-            if isinstance(element, DugElement):
-                element.add_concept(self.concepts[identifier.id])
-
-            # If element is actually a user defined concept (e.g. TOPMedTag), associate ident with concept
-            # Child elements of these user-defined concepts will inherit all these identifiers as well.
-            elif isinstance(element, DugConcept):
-                element.add_identifier(identifier)
-
-    def expand_concept(self, concept):
-
-        # Get knowledge graphs of terms related to each identifier
-        for ident_id, identifier in concept.identifiers.items():
-
-            # Conditionally skip some identifiers if they are listed in config
-            if ident_id in self.exclude_identifiers:
-                continue
-
-            # Use pre-defined queries to search for related knowledge graphs that include the identifier
-            for query_name, query_factory in self.tranql_queries.items():
-
-                # Skip query if the identifier is not a valid query for the query class
-                if not query_factory.is_valid_curie(ident_id):
-                    logger.info(f"identifier {ident_id} is not valid for query type {query_name}. Skipping!")
-                    continue
-
-                # Fetch kg and answer
-                kg_outfile = f"{self.crawlspace}/{ident_id}_{query_name}.json"
-                answers = self.tranqlizer.expand_identifier(ident_id, query_factory, kg_outfile)
-
-                # Add any answer knowledge graphs to
-                for answer in answers:
-                    concept.add_kg_answer(answer, query_name=query_name)
-
-
-def get_parser(hook, parser_name):
-    """Get the parser from all parsers registered via the define_parsers hook"""
-
-    available_parsers = {}
-    hook.define_parsers(parser_dict=available_parsers)
-    parser = available_parsers.get(parser_name.lower())
-    if parser is not None:
-        return parser
-
-    err_msg = f"Cannot find parser of type '{parser_name}'\n" \
-              f"Supported parsers: {', '.join(available_parsers.keys())}"
-    logger.error(err_msg)
-    raise ParserNotFoundException(err_msg)
-
-
-def get_targets(target_name) -> Iterable[Path]:
-    if target_name.startswith("http://") or target_name.startswith("https://"):
-        loader = partial(load_from_network, os.getenv("DUG_DATA_DIR", "data"))
-    else:
-        loader = load_from_filesystem
-    return loader(target_name)
-
-
-class Dug:
-    concepts_index = "concepts_index"
-    variables_index = "variables_index"
-    kg_index = "kg_index"
-
-    def __init__(self, cfg: Config):
-        self._cfg = cfg
-        self._search = Search(cfg=self._cfg, indices=[self.concepts_index, self.variables_index, self.kg_index])
-
-    def crawl(self, target_name: str, parser_type: str, element_type: str = None):
-
-        pm = get_plugin_manager()
-        parser = get_parser(pm.hook, parser_type)
-        targets = get_targets(target_name)
-
-        for target in targets:
-            self._crawl(target, parser, element_type)
-
-    def _crawl(self, target: Path, parser: Parser, element_type):
-
-        # Configure redis so we can fetch things from cache when needed
-        redis_connection = redis.StrictRedis(host=self._cfg.redis_host,
-                                             port=self._cfg.redis_port,
-                                             password=self._cfg.redis_password)
-
-        http_session = CachedSession(cache_name='annotator',
-                                     backend='redis',
-                                     connection=redis_connection)
-
-        # Create annotation engine for fetching ontology terms based on element text
-        preprocessor = anno.Preprocessor(**self._cfg.preprocessor)
-        annotator = anno.Annotator(**self._cfg.annotator)
-        normalizer = anno.Normalizer(**self._cfg.normalizer)
-        synonym_finder = anno.SynonymFinder(**self._cfg.synonym_service)
-        ontology_helper = anno.OntologyHelper(**self._cfg.ontology_helper)
-        tranqlizer = anno.ConceptExpander(**self._cfg.concept_expander)
-
-        # Greenlist of ontology identifiers that can fail normalization and still be valid
-        ontology_greenlist = self._cfg.ontology_greenlist if hasattr(self._cfg, "ontology_greenlist") else []
-
-        # DugAnnotator combines all annotation components into single annotator
-        dug_annotator = anno.DugAnnotator(preprocessor=preprocessor,
-                                          annotator=annotator,
-                                          normalizer=normalizer,
-                                          synonym_finder=synonym_finder,
-                                          ontology_helper=ontology_helper,
-                                          ontology_greenlist=ontology_greenlist)
-
-        # Initialize crawler
-        crawler = Crawler(crawl_file=str(target),
-                          parser=parser,
-                          annotator=dug_annotator,
-                          tranqlizer=tranqlizer,
-                          tranql_queries=self._cfg.tranql_queries,
-                          http_session=http_session,
-                          exclude_identifiers=self._cfg.tranql_exclude_identifiers,
-                          element_type=element_type)
-
-        # Read elements, annotate, and expand using tranql queries
-        crawler.crawl()
-
-        # Index Annotated Elements
-        for element in crawler.elements:
-            # Only index DugElements as concepts will be indexed differently in next step
-            if not isinstance(element, DugConcept):
-                self._search.index_element(element, index=self.variables_index)
-
-        # Index Annotated/TranQLized Concepts and associated knowledge graphs
-        for concept_id, concept in crawler.concepts.items():
-            self._search.index_concept(concept, index=self.concepts_index)
-
-            # Index knowledge graph answers for each concept
-            for kg_answer_id, kg_answer in concept.kg_answers.items():
-                self._search.index_kg_answer(concept_id=concept_id,
-                                             kg_answer=kg_answer,
-                                             index=self.kg_index,
-                                             id_suffix=kg_answer_id)
-
-    def search(self, target, query, **kwargs):
-        targets = {
-            'concepts': partial(
-                self._search.search_concepts, index=kwargs.get('index', self.concepts_index)),
-            'variables': partial(
-                self._search.search_variables, index=kwargs.get('index', self.variables_index), concept=kwargs.pop('concept', None)),
-            'kg': partial(
-                self._search.search_kg, index=kwargs.get('index', self.kg_index), unique_id=kwargs.pop('unique_id', None)),
-            'nboost': partial(
-                self._search.search_nboost, index=kwargs.get('index', None)),
-        }
-        kwargs.pop('index', None)
-        func = targets.get(target)
-        if func is None:
-            raise ValueError(f"Target must be one of {', '.join(targets.keys())}")
-
-        return func(query=query, **kwargs)
-
-    def status(self):
-        ...
+class SearchException(Exception):
+    def __init__(self, message, details):
+        self.message = message
+        self.details = details
