+++ tests/integration/test_crawler.py
@@ -2,7 +2,7 @@ import tempfile
 from pathlib import Path
 from unittest.mock import MagicMock
 
-from dug.core import Crawler
+from dug.core.crawler import Crawler
 from tests.integration.conftest import TEST_DATA_DIR
