+++ tests/unit/test_annotate.py
@@ -4,7 +4,7 @@ from typing import List
 import pytest
 
 from dug.config import Config
-from dug.annotate import Identifier, Preprocessor, Annotator, Normalizer, SynonymFinder, OntologyHelper
+from dug.core.annotate import Identifier, Preprocessor, Annotator, Normalizer, SynonymFinder, OntologyHelper
 
 
 def test_identifier():
